finalized avns

+++ avns/greedy_insert.py
@@ -21,7 +21,7 @@ def try_inserting_to_vehicles(solution:Solution, sorted_custs_idx: np.ndarray)->
             item_dims: np.ndarray = np.zeros([total_num_items, 3], dtype=float)
             item_volumes: np.ndarray = np.zeros([total_num_items, ], dtype=float)
             item_weights: np.ndarray = np.zeros([total_num_items, ], dtype=float)
-            item_priorities: np.ndarray = np.zeros([total_num_items, ], dtype=float)
+            item_priorities: np.ndarray = np.zeros([total_num_items, ], dtype=int)
             n = 0
             for i, cust_idx in enumerate(route):
                 c_num_items = solution.node_num_items[cust_idx]
